Fifth Commit

+++ projectProductApplication/src/main/java/com/projectApplication/repository/productRepositortory/ProductSkuEntityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductSkuEntityRepository extends JpaRepository<ProductSkuEntity, Integer> {
+public interface ProductSkuEntityRepository extends JpaRepository<ProductSkuEntity, Long> {
 }
